DAOS-4452 vos: minor epoch for incarnation log
Introduce 16-bits minor epoch for the modifications against the
same target (object/dkey/akey) via the same DTX that will share
the same incarnation log entry.

It is possible that a distributed transaction may contains both
punch and update against the same object/dkey/akey. So the ilog
entry may change between 'punch' and 'update' status frequently.

To support that, we add two minor epochs to an ilog entry, one
is for punch, another is for update. When the incarnation log
entry is created for the first time or re-created after punch,
its update minor epoch is set as current minor epoch (that is
controlled by the DTX handle); similarly, when it is punched,
its punch minor is set as current minor epoch. For a give ilog
entry, its minor epoch within the same DTX is increased only,
so by comparing the punch minor epoch and the update one, we
can know whether the ilog entry is a punched one or update.

The two minor epochs are stored inside btr_record::rec_hkey,
that is reserved space for the ilog entry record key, so not
consume additional space.

On the other hand, the minor epoch is only visible inside VOS,
it does not participate in ilog_hkey_cmp(), so will not affect
the ilog entry's visibility that is still controlled by main
epoch as original case does.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>