CORCI-918 build: Refactor Jenkinsfile
This is part of a major refactoring of the Jenkinsfile to reduce the size
of the code and hopefully make it easier to maintain.

This commit is mostly moving inline scripts into the DAOS repository,
and starting to use the STAGE_NAME environment variable to control
what the stage does.

Signed-off-by: John E Malmberg <john.e.malmberg@intel.com>